Rename counterActions -> counter.
